correct channel uint8

+++ load_functions/prepare_split_images.py
@@ -70,6 +70,7 @@ def diplay_img_info(img, divisor, use_RGB):
         nr_channels = 1
     return nr_z_slices, nr_channels, nr_timepoints, x_dim, y_dim, x_div, y_div 
 
+
 def correct_channels(img):
   '''For 2D + T (with or without RGB) a artificial z channel gets created'''
   if img.shape[-1] ==3:
@@ -78,12 +79,12 @@ def correct_channels(img):
     use_RGB = False
   if len(img.shape) ==4 and use_RGB:
     t, x, y, c = img.shape
-    zeros = np.zeros((t,1,y,x,c))
+    zeros = np.zeros((t,1,y,x,c), dtype=np.uint8 )
     zeros[:,0,:,:,:] = img
     img = zeros
   elif len(img.shape) ==3 and not use_RGB:
     t, x, y = img.shape
-    zeros = np.zeros((t,1,y,x))
+    zeros = np.zeros((t,1,y,x), dtype=np.uint8 )
     zeros[:,0,:,:] = img
     img = zeros
   return img, use_RGB
